[REFACTOR] 폴더 이름 수정으로 인한 경로 변경
