+++ frontend/src/views/review/ReviewCreatePage.vue
@@ -1,0 +1,19 @@
+<template>
+    <ReviewPageContent>
+
+    </ReviewPageContent>
+</template>
+
+<script>
+import ReviewPageContent from '@/components/WhiteContentBox.vue';
+
+export default{
+    components: {
+        ReviewPageContent
+    }
+}
+</script>
+
+<style scoped>
+
+</style>
